(change): split default featuremap
