+++ qxmt/feature_maps/pennylane/__init__.py
@@ -1,15 +1,10 @@
-from qxmt.feature_maps.pennylane.defaults import (
-    HRotationFeatureMap,
-    RotationFeatureMap,
-    XXFeatureMap,
-    YYFeatureMap,
-    ZZFeatureMap,
-)
+from qxmt.feature_maps.pennylane.ising import XXFeatureMap, YYFeatureMap, ZZFeatureMap
+from qxmt.feature_maps.pennylane.rotation import HRotationFeatureMap, RotationFeatureMap
 
 __all__ = [
-    "HRotationFeatureMap",
-    "RotationFeatureMap",
     "XXFeatureMap",
     "YYFeatureMap",
     "ZZFeatureMap",
+    "HRotationFeatureMap",
+    "RotationFeatureMap",
 ]
