+++ tests/feature_maps/pennylane/test_imports.py
@@ -1,11 +1,11 @@
 from qxmt.feature_maps.pennylane import __all__
 
 EXPECTED_ALL = [
-    "HRotationFeatureMap",
-    "RotationFeatureMap",
     "XXFeatureMap",
     "YYFeatureMap",
     "ZZFeatureMap",
+    "HRotationFeatureMap",
+    "RotationFeatureMap",
 ]
